created Box.cs
Box script contains list of ingredients (public for now as we are hard coding things) as well as a propert to change whether or not the box is open. If the box is open. then it will display a gui to the bottem right corner (hopefully) showing what ingrediants it contains